requirements.txt
Flask==3.1.2
mongoengine==0.29.1
Werkzeug==3.1.3
flask-cors==6.0.1
pymongo==4.15.3
python-dotenv==1.1.1
Flask-JWT-Extended==4.7.1


waitress==3.0.2

